+++ CHANGELOG.md
@@ -14,6 +14,7 @@ since `1.7.0`.
   [#13](https://github.com/libtcod/libtcod/issues/13)
   [#50](https://github.com/libtcod/libtcod/pull/50)
 - The `TCOD_VSYNC` environment variable was being ignored.
+- Fixed characters being dropped when color codes were used.
 
 ## [1.16.0-alpha.8] - 2020-04-03
 ### Fixed